+++ src/clustering/record-linkage/key-collision.js
@@ -7,6 +7,9 @@
  * keys collide.
  */
 import RecordLinkageClusterer from './abstract';
+import {
+  clustersFromSetGraph
+} from './helpers';
 
 /**
  * Key Collision Clusterer class.
@@ -19,12 +22,13 @@ export class KeyCollisionClusterer extends RecordLinkageClusterer {
 
     // Validating keyer
     this.keyer = params.keys || params.key;
+    this.merge = params.merge || false;
 
     if (typeof this.keyer !== 'function')
       throw new Error('talisman/clustering/record-linkage/key-collision: the given keyer is not a function.');
   }
 
-  run() {
+  runWithMerge() {
     const map = Object.create(null);
 
     // Computing buckets map
@@ -35,7 +39,53 @@ export class KeyCollisionClusterer extends RecordLinkageClusterer {
       for (let j = 0, m = keys.length; j < m; j++) {
         const key = keys[j];
 
-        // If the key is falsey, we continue
+        // If the key is falsy, we continue
+        if (!key)
+          continue;
+
+        if (!map[key])
+          map[key] = new Set();
+        map[key].add(i);
+      }
+    }
+
+    // Computing graph
+    // TODO: I sense that we can do better & faster
+    const graph = Object.create(null);
+
+    for (const key in map) {
+      const bucket = Array.from(map[key]);
+
+      for (let i = 0, l = bucket.length; i < l; i++) {
+        for (let j = i + 1; j < l; j++) {
+          graph[bucket[i]] = graph[bucket[i]] || new Set();
+          graph[bucket[i]].add(bucket[j]);
+
+          graph[bucket[j]] = graph[bucket[j]] || new Set();
+          graph[bucket[j]].add(bucket[i]);
+        }
+      }
+    }
+
+    return clustersFromSetGraph(
+      this.items,
+      graph,
+      this.params.minClusterSize
+    );
+  }
+
+  runWithoutMerge() {
+    const map = Object.create(null);
+
+    // Computing buckets map
+    for (let i = 0, l = this.items.length; i < l; i++) {
+      const item = this.items[i],
+            keys = [].concat(this.keyer(item, i));
+
+      for (let j = 0, m = keys.length; j < m; j++) {
+        const key = keys[j];
+
+        // If the key is falsy, we continue
         if (!key)
           continue;
 
@@ -54,6 +104,13 @@ export class KeyCollisionClusterer extends RecordLinkageClusterer {
     }
 
     return clusters;
+  }
+
+  run() {
+    if (this.merge)
+      return this.runWithMerge();
+    else
+      return this.runWithoutMerge();
   }
 }